+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## not released
 
+- Fix: Weapons were not removed for upgrades if they are stored as plural
+
 ## v0.6.3 (2023-05-19)
 
 - Fix: Output spells only for army with psychic