+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fix: Weapons were not removed for upgrades if they are stored as plural
 - Fix: #3 Missing equipment in rule overview
 - Add: Show W6 for knocked out test
+- Add: List overview
 
 ## v0.6.3 (2023-05-19)